Lots of changes were made!
Hello, I won't make an introduction but there are quite a few
changes to this repo. I haven't touched this project in almost
a year. I only brought it back after playing with the battle sim.

Most of the work done is explained in the class files and class
info file. It is mainly structural work. I made classes for
different aspects of the game that need to store information.

Players obviously have a lot of info to maintain. Cards, money, etc.
The board is a graph where each node is a territory.
Territories house a lot of info, from its value to who is on it.

The game class is not complete but I leave it here for record keeping.
There are several test files to ensure the classes work as expected.
I did not include them because they are not integral to the project.

The new text file is the representation of the risk europe board
in my board scripting lang (if you want to call it that).
I think it is read only lol, but it creates a functionally identical
game board to risk europe.

I know it is not good practice to make long logs. I do it anyways
because there is a lot and I haven't updated this in a year.

Chill out for christ's sake!