Bid system created + minor changes
The bid sequence was created. It replicates every aspect of the
bidding sequence in the board game. The only change is that ties
can only occur once. Instead of rolling one die potentially infinite
times, two dice will be rolled once with the winner being
whoever is last in the player array.

Some helper methods were created in the game class.
One is to print new lines to clear the screen.
The other is an easy way to get input from the user within a range.
They will be used extensively since this is a CLI program.

The army class had a small typo in the split method.