Overhauled pickCards() and useCard() bug
pickCards(i,i) has now become pickCard(i)
Instead of picking two cards at once only one card is picked.
This was changed to assist the Game class in getting input.
It also makes the system able to handle odd numbers of active cards
if that ever becomes a possibility.

useCard() had a bug where if it picked a null card it would still
add that to the discard pile which will eventually cause the loss
of valid card ids in drawPile when it gets reset prematurely.