Added card picking procedure + more
assignOrder() places the firstPlayer reference first in the array.
It rotates players leftwards until the firstPlayer is in place.

pickingSequence() is the method that interfaces with the user to
choose cards. This implementation needed some changes in the Player
class. Namely picking cards needed to be done one at a time. This
made it much easier to write and it made more sense from a class
design standpoint to only pick one instead of two at a time.

A huge bug was squashed in the giveCards() method in Player.
It would assign the reference of the cardId array passed instead of
copying the data. This resulted in each player sharing the same drawPile
as they were all intialized with the same array reference. Yikes.