Added setup sequence + many other changes
A variety of addtional methods and data fields were added to the
Player, Army, Board, and Territory classes. The territory picking
sequence was completed. The structure of the turn system is
commented out.

Territories that are cities now can be marked as a potential
starting location.

The loadFromFile() method in Board needed to be updated for the
new canStart field. The risk_eu_board file implements the new syntax.

The Player class saw an addtion of a HashMap to keep track of
units and their locations. A new method was added to add new units.
There will be more in the future.

Ignore class-layout.txt as it is severly outdated.