Added win condition
The game is now complete. Meaning you can play through and finish
a game. Ties are settled with money instead of territory count.

More testing will be done to fix any bugs. Even though the game is
complete there is still plenty to work on. Bonuses still need to
be added.